немного фикса
